Rework post components in some small coponents
